refactor: 공용 코멘트 컴포넌트 위치 변경
